+++ HISTORY.rst
@@ -6,9 +6,10 @@ History
 0.2.7 (unreleased)
 ------------------
 
-* Fix issues #94 #97 #92 (?)
+* Fix issues #94 #97 #92 #100
 * Add support for reading per-exposure flux calibration and correction vectors.
 * Add plot functions for per-fiber data vs fiber number or focal-plane position.
+* Add a plug_map attribute to spPlate, spFrame, spCFrame.
 
 0.2.6 (2015-08-05)
 ------------------